requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-monitor==6.0.2
boto3==1.34.149
botocore==1.34.149
colorama==0.4.6
jmespath==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.2